Settings: Rename delete-site-options component to site-tools
To reflect title and purpose.
